big bug in Obj.add_offset
git-svn-id: http://caml.inria.fr/svn/ocaml/version/3.11@9559 f963ae5c-01c2-4b8c-9fe0-0dff7051ff02

+++ byterun/obj.c
@@ -167,7 +167,7 @@ CAMLprim value caml_obj_truncate (value v, value newsize)
 
 CAMLprim value caml_obj_add_offset (value v, value offset)
 {
-  return v + Int32_val (offset);
+  return v + (unsigned long) Int32_val (offset);
 }
 
 /* The following functions are used in stdlib/lazy.ml.
